Rename api -> rest-api
